test: remove old test, add start of node tests

+++ tests/integration_tests/conftest.py
@@ -1,9 +1,15 @@
+# pylint: disable=missing-timeout
 from collections.abc import Callable
 
 import pytest
 import requests
+import structlog
 
 from tests.integration_tests.helpers import API_URL
+
+logger = structlog.get_logger(__name__)
+
+TEST_CLEANUP = False
 
 
 @pytest.fixture
@@ -11,7 +17,9 @@ def products() -> dict[str, str]:
     """Return mapping from product name to product ids"""
     products_response = requests.get(f"{API_URL}/products")
     assert products_response.ok
-    return {product["name"]: product["product_id"] for product in (products_response.json())}
+    return {
+        product["name"]: product["product_id"] for product in (products_response.json())
+    }
 
 
 @pytest.fixture
@@ -40,10 +48,13 @@ def get_subscription_processes():
 def clean(get_subscriptions, get_subscription_processes):
     yield
 
-    for sub_id, details in get_subscriptions().items():
-        processes = get_subscription_processes(sub_id)
-        for pid, _ in processes.items():
-            assert requests.delete(f"{API_URL}/processes/{pid}").ok
+    if TEST_CLEANUP:
+        for sub_id, details in get_subscriptions().items():
+            processes = get_subscription_processes(sub_id)
+            for pid, _ in processes.items():
+                assert requests.delete(f"{API_URL}/processes/{pid}").ok
 
-        response = requests.delete(f"{API_URL}/subscriptions/{sub_id}")
-        assert response.ok, f"Failed to delete subscription {sub_id} ({details['description']}): {response.text}"
+            response = requests.delete(f"{API_URL}/subscriptions/{sub_id}")
+            assert (
+                response.ok
+            ), f"Failed to delete subscription {sub_id} ({details['description']}): {response.text}"
